[ANDROSDK-1964] Move CategoryComboFields to network package
