+++ core/src/main/java/org/hisp/dhis/android/network/category/CategoryComboNetworkHandlerImpl.kt
@@ -27,12 +27,10 @@
  */
 package org.hisp.dhis.android.network.category
 
-import org.hisp.dhis.android.core.arch.api.fields.internal.Fields
-import org.hisp.dhis.android.core.arch.api.filters.internal.Filter
+import org.hisp.dhis.android.core.arch.api.payload.internal.Payload
 import org.hisp.dhis.android.core.category.CategoryCombo
 import org.hisp.dhis.android.core.category.internal.CategoryComboNetworkHandler
 import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
-import org.hisp.dhis.android.network.common.PayloadJson
 import org.koin.core.annotation.Singleton
 
 @Singleton
@@ -41,12 +39,12 @@ internal class CategoryComboNetworkHandlerImpl(
 ) : CategoryComboNetworkHandler {
     private val service: CategoryComboService = CategoryComboService(httpClient)
 
-    override suspend fun getCategoryCombos(
-        fields: Fields<CategoryCombo>,
-        uids: Filter<CategoryCombo>,
-        paging: Boolean,
-    ): PayloadJson<CategoryCombo> {
-        val apiPayload = service.getCategoryCombos(fields, uids, paging)
+    override suspend fun getCategoryCombos(categoryComboUids: Set<String>): Payload<CategoryCombo> {
+        val apiPayload = service.getCategoryCombos(
+            CategoryComboFields.allFields,
+            CategoryComboFields.uid.`in`(categoryComboUids),
+            paging = false,
+        )
         return apiPayload.mapItems(::categoryComboDtoToDomainMapper)
     }
 }
